\name{MyOwnShark-package}
\alias{MyOwnShark-package}
\alias{MyOwnShark}
\docType{package}
\title{
\packageTitle{MyOwnShark}
}
\description{
\packageDescription{MyOwnShark}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MyOwnShark}
\packageIndices{MyOwnShark}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{MyOwnShark}

Maintainer: \packageMaintainer{MyOwnShark}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
